+++ tests/conftest.py
@@ -1,4 +1,4 @@
-from os import environ, pardir, path
+from os import environ
 from pathlib import Path
 
 import pytest
@@ -6,15 +6,21 @@ import pytest
 PYTESTS_DIR = Path(__file__).parent
 PROJECT_DIR = PYTESTS_DIR.parent
 LOG_DIR = PROJECT_DIR.joinpath('log').resolve()
-FIXTURES_DIR = PYTESTS_DIR.joinpath('file_fixtures')
-RENDERED_FIXTURES_DIR = FIXTURES_DIR.joinpath('rendered')
+TMP_DIR = PYTESTS_DIR.joinpath('tmp')
 
 # Some env vars that we need or are helpful for pytest
 environ['INVOKED_BY_PYTEST'] = 'True'
 environ['YARALYZER_LOG_DIR'] = str(LOG_DIR)
 
+from yaralyzer.helpers.env_helper import is_env_var_set_and_not_false     # noqa: E402
 from yaralyzer.helpers.file_helper import files_in_dir, load_binary_data  # noqa: E402
 from yaralyzer.yaralyzer import Yaralyzer                                 # noqa: E402
+
+FIXTURES_DIR = PYTESTS_DIR.joinpath('file_fixtures')
+YARA_FIXTURES_DIR = FIXTURES_DIR.joinpath('yara_rules')
+RENDERED_FIXTURES_DIR = FIXTURES_DIR.joinpath('rendered')
+PYTEST_REBUILD_FIXTURES_ENV_VAR = 'PYTEST_REBUILD_FIXTURES'
+SHOULD_REBUILD_FIXTURES = is_env_var_set_and_not_false(PYTEST_REBUILD_FIXTURES_ENV_VAR)
 
 
 # Full paths to file fixtures
@@ -25,7 +31,7 @@ def il_tulipano_path() -> Path:
 
 @pytest.fixture(scope='session')
 def tulips_yara_path() -> Path:
-    return FIXTURES_DIR.joinpath('yara_rules', 'tulips.yara')
+    return YARA_FIXTURES_DIR.joinpath('tulips.yara')
 
 
 @pytest.fixture(scope='session')
@@ -39,7 +45,7 @@ def binary_file_bytes(binary_file_path) -> bytes:
 
 
 @pytest.fixture(scope='session')
-def tulips_yara_regex() -> str:
+def tulips_yara_pattern() -> str:
     """The regex that appears in the tulips_yara_path() fixture"""
     return 'tulip.{1,2500}tulip'
 
@@ -53,9 +59,7 @@ def a_yaralyzer(il_tulipano_path, tulips_yara_path) -> Yaralyzer:
 @pytest.fixture
 def tmp_dir() -> Path:
     """Clear the tmp dir when fixture is loaded"""
-    tmpdir = PYTESTS_DIR.joinpath('tmp')
-
-    for tmp_file in files_in_dir(tmpdir):
+    for tmp_file in files_in_dir(TMP_DIR):
         tmp_file.unlink()
 
-    return tmpdir
+    return TMP_DIR